refactor: seperate files for resepective ops
