+++ pkg/gofr/cron.go
@@ -2,11 +2,6 @@ package gofr
 
 import (
 	"context"
-	"errors"
-	"fmt"
-	"regexp"
-	"strconv"
-	"strings"
 	"sync"
 	"time"
 
@@ -76,203 +71,6 @@ func NewCron(cntnr *container.Container) *Crontab {
 	}()
 
 	return c
-}
-
-// this will compile the regex once instead of compiling it each time when it is being called.
-var (
-	matchSpaces = regexp.MustCompile(`\s+`)
-	matchN      = regexp.MustCompile(`(.*)/(\d+)`)
-	matchRange  = regexp.MustCompile(`^(\d+)-(\d+)$`)
-)
-
-// parseSchedule parses schedule string and create job struct with filled times to launch,
-// or error if syntax is wrong.
-func parseSchedule(s string) (*job, error) {
-	var err error
-
-	j := &job{}
-	s = matchSpaces.ReplaceAllLiteralString(s, " ")
-	s = strings.Trim(s, " ")
-	parts := strings.Split(s, " ")
-
-	var partsItr int
-
-	switch len(parts) {
-	case schedulePartsWithSecond:
-		j.sec, err = parsePart(parts[partsItr], 0, seconds)
-		if err != nil {
-			return nil, err
-		}
-
-		partsItr++
-	case scheduleParts:
-		partsItr = 0
-	default:
-		return nil, errBadScheduleFormat
-	}
-
-	j.min, err = parsePart(parts[partsItr], 0, minutes)
-	if err != nil {
-		return nil, err
-	}
-
-	j.hour, err = parsePart(parts[partsItr+1], 0, hrs)
-	if err != nil {
-		return nil, err
-	}
-
-	j.day, err = parsePart(parts[partsItr+2], 1, days)
-	if err != nil {
-		return nil, err
-	}
-
-	j.month, err = parsePart(parts[partsItr+3], 1, months)
-	if err != nil {
-		return nil, err
-	}
-
-	j.dayOfWeek, err = parsePart(parts[partsItr+4], 0, dayOfWeek)
-	if err != nil {
-		return nil, err
-	}
-
-	//  day/dayOfWeek combination
-	mergeDays(j)
-
-	return j, nil
-}
-
-func mergeDays(j *job) {
-	switch {
-	case len(j.day) < 31 && len(j.dayOfWeek) == 7: // day set, but not dayOfWeek, clear dayOfWeek
-		j.dayOfWeek = make(map[int]struct{})
-	case len(j.dayOfWeek) < 7 && len(j.day) == 31: // dayOfWeek set, but not day, clear day
-		j.day = make(map[int]struct{})
-	}
-}
-
-// parsePart parse individual schedule part from schedule string.
-func parsePart(s string, minValue, maxValue int) (map[int]struct{}, error) {
-	// wildcard pattern
-	if s == "*" {
-		return getDefaultJobField(minValue, maxValue, 1), nil
-	}
-
-	// */2 1-59/5 pattern
-	if matches := matchN.FindStringSubmatch(s); matches != nil {
-		return parseSteps(s, matches[1], matches[2], minValue, maxValue)
-	}
-
-	// 1,2,4 or 1,2,10-15,20,30-45 pattern
-	return parseRange(s, minValue, maxValue)
-}
-
-func parseSteps(s, match1, match2 string, minValue, maxValue int) (map[int]struct{}, error) {
-	localMin := minValue
-	localMax := maxValue
-
-	if match1 != "" && match1 != "*" {
-		rng := matchRange.FindStringSubmatch(match1)
-		if rng == nil {
-			return nil, errParsing{match1, s}
-		}
-
-		localMin, _ = strconv.Atoi(rng[1])
-		localMax, _ = strconv.Atoi(rng[2])
-
-		if localMin < minValue || localMax > maxValue {
-			return nil, errOutOfRange{rng[1], s, minValue, maxValue}
-		}
-	}
-
-	n, _ := strconv.Atoi(match2)
-
-	return getDefaultJobField(localMin, localMax, n), nil
-}
-
-func parseRange(s string, minValue, maxValue int) (map[int]struct{}, error) {
-	r := make(map[int]struct{})
-	parts := strings.Split(s, ",")
-
-	for _, part := range parts {
-		if err := processPart(part, minValue, maxValue, r); err != nil {
-			return nil, err
-		}
-	}
-
-	if len(r) == 0 {
-		return nil, errParsing{invalidPart: s}
-	}
-
-	return r, nil
-}
-
-func processPart(part string, minValue, maxValue int, r map[int]struct{}) error {
-	if rng := matchRange.FindStringSubmatch(part); rng != nil {
-		return handleRange(rng, minValue, maxValue, r)
-	}
-	return handleValue(part, minValue, maxValue, r)
-}
-
-func handleValue(value string, minValue, maxValue int, r map[int]struct{}) error {
-	i, err := strconv.Atoi(value)
-	if err != nil {
-		return errParsing{invalidPart: value}
-	}
-
-	if i < minValue || i > maxValue {
-		return errOutOfRange{
-			rangeVal: i,
-			input:    value,
-			min:      minValue,
-			max:      maxValue,
-		}
-	}
-
-	r[i] = struct{}{}
-	return nil
-}
-
-func handleRange(rng []string, minValue, maxValue int, r map[int]struct{}) error {
-	localMin, errMin := strconv.Atoi(rng[1])
-	localMax, errMax := strconv.Atoi(rng[2])
-
-	if errMin != nil || errMax != nil {
-		return fmt.Errorf("invalid range format: %v", rng)
-	}
-
-	if localMin > localMax {
-		return errOutOfRange{
-			rangeVal: fmt.Sprintf("%d-%d", localMin, localMax),
-			input:    fmt.Sprintf("%d-%d", localMin, localMax),
-			min:      minValue,
-			max:      maxValue,
-		}
-	}
-
-	if localMin < minValue || localMax > maxValue {
-		return errOutOfRange{
-			rangeVal: fmt.Sprintf("%d-%d", localMin, localMax),
-			input:    fmt.Sprintf("%d-%d", localMin, localMax),
-			min:      minValue,
-			max:      maxValue,
-		}
-	}
-
-	for i := localMin; i <= localMax; i++ {
-		r[i] = struct{}{}
-	}
-	return nil
-}
-
-func getDefaultJobField(minValue, maxValue, incr int) map[int]struct{} {
-	r := make(map[int]struct{})
-
-	for i := minValue; i <= maxValue; i += incr {
-		r[i] = struct{}{}
-	}
-
-	return r
 }
 
 func (c *Crontab) runScheduled(t time.Time) {
@@ -364,8 +162,6 @@ func (c *Crontab) AddJob(schedule, jobName string, fn CronFunc) error {
 
 	return nil
 }
-
-var errBadScheduleFormat = errors.New("schedule string must have five components like * * * * *")
 
 // noopRequest is a non-operating implementation of Request interface
 // this is required to prevent panics while executing cron jobs.
